backend/TEST-FOLDER/requirements.txt
crewai==0.41.1
crewai-tools==0.8.3
pandas==2.0.3
openpyxl==3.1.2
GitPython==3.1.40
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0